Fixed quantity check for appointment creation and clears cart afterwards.
Fixed the issue where when making an appointment, it would incorrectly fail due to a "lack of items" when there are items available.  Added a cart clear functionality after appointment is created.  Added a clear_cart function to cart_item DAOs that clears all cart items of a given account id.